requirements.txt
pandas
Pillow
tensorboard
optuna
torchvision
matplotlib
seaborn
statsmodels
datasets
